+++ neo-compiler.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neon", "neon\neon.csproj", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "neoj", "neoj\neoj.csproj", "{5FB7A952-6EAC-4BF9-AD23-AF957EF21925}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "neon-lib", "neon-lib\neon-lib.csproj", "{74329CAE-F53D-4820-8982-0E873439916F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{5FB7A952-6EAC-4BF9-AD23-AF957EF21925}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FB7A952-6EAC-4BF9-AD23-AF957EF21925}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5FB7A952-6EAC-4BF9-AD23-AF957EF21925}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74329CAE-F53D-4820-8982-0E873439916F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74329CAE-F53D-4820-8982-0E873439916F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74329CAE-F53D-4820-8982-0E873439916F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74329CAE-F53D-4820-8982-0E873439916F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
