+++ neon/neon.csproj
@@ -29,8 +29,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Mono.Cecil" Version="0.10.0" />
-    <PackageReference Include="Neo.VM" Version="2.3.0" />
+    <ProjectReference Include="..\neon-lib\neon-lib.csproj" />
   </ItemGroup>
-
+  <ItemGroup>
+    <Compile Remove="MSIL\CctorSubVM.cs" />
+    <Compile Remove="MSIL\Conv_Common.cs" />
+    <Compile Remove="MSIL\ILModule.cs" />
+  </ItemGroup>
 </Project>
